create new folder for pages, move home page there
